:truck: Extract styled components
