+++ src/components/List/Ul.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const Ul = styled.ul`
+  margin-top: 2rem;
+  margin-bottom: 1.5rem;
+  list-style: none;
+  overflow: auto;
+  max-height: calc(100vh - 500px);
+`;
+
+export default Ul;
