+++ src/components/ListItem/DisplayText.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const DisplayText = styled.div`
+  width: calc(100% - 4rem);
+  display: inline-block;
+  outline: none;
+  letter-spacing: normal;
+`;
+
+export default DisplayText;
