+++ src/components/ListItem/InputText.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const InputText = styled.input`
+  width: calc(100% - 4rem);
+  display: inline-block;
+  outline: none;
+  letter-spacing: normal;
+`;
+
+
+export default InputText;
